More CRUD Work, almost got the image uploading down
Id say user image uploading is a next phase thing. we just need them to
display properly on the site.  so we can do that in hard put divs or
whatever. next step for me will be getting an orders tab working and a
homepage so we can start displaying stuff.